TMDM-12197 Let keep original message in before saving&deleting,entity process, refactor code, add unit test.
